Devious Contraptions Struggle Minigame by Frayed
A minigame to offer gameplay value to being locked into contraptions.

This version has a single minigame that replaces the default "struggle" escape option. The "pick lock" and "break device" escape options are unchanged. Basic instructions on how to play the minigame are included in-game as a tutorial message whenever you start it.

Escape minigame:
The basic idea of the minigame is that to free themselves, the player has to hold a specific sequence of movement keys for at least a certain amount of time. If a player hits the wrong key or doesn't hold it long enough, the entered sequence resets and they have to start over. E.g. if the sequence is "WDSA", if the player first hits "W" they will get FX to indicate the first guess is correct. If they then hit "S" they get FX to indicate their guess for the second key is wrong (a "fail"). They then have to start from the first key again, which if they paid attention they should know is "W".

There is an optional chance to allow for a "critical fail". This is a chance that on a fail a new sequence to guess is rolled, e.g. if it was "WDSA", it could become "SDWA". All previously memorized keys are then useless and the player has to start over. A critical fail has its own FX to indicate to the player that it happened.

A further difficulty setting is "escalation". If enabled, on a critical fail there is an additional chance that the new sequence is longer than the previous one, e.g. if it was "WDSA", it could become "SADWA". This introduces some risk to the game: wrong guesses or just bad luck can make the game much harder.

All these difficulty options can be controlled via the DD MCM.

Certain parts of the minigame script are exposed as an API for mod authors to interact with. Difficulty options can be set and locked via script by a mod author calling SetDifficultyParameters. The progress of the player in the minigame can be tracked via script or via Conditional variables, to facilitate e.g. NPC comments or actions.